Unity: Implemented moveTo and faceToward actions and added some debug code to record images over time to analyze as a time sequence video.

+++ unity/RoBart/Assets/Scripts/Actions/Actions.cs
@@ -10,7 +10,7 @@ public struct MoveAction
 
 public struct MoveToAction
 {
-    public float pointNumber;
+    public int pointNumber;
 }
 
 public struct TurnInPlaceAction
@@ -20,7 +20,7 @@ public struct TurnInPlaceAction
 
 public struct FaceTowardAction
 {
-    public float pointNumber;
+    public int pointNumber;
 }
 
 public struct FaceTowardHeadingAction
